use npm init to create project
closes #51
